backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "agent-testing-backend"
version = "0.1.0"
description = "Agent Testing Platform - Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "pydantic>=2.0",
    "pydantic-settings",
    "pyjwt==2.12.1",
    "python-socketio",
    "redis",
    "celery[redis]",
    "livekit-agents>=1.2.6",
    "livekit-plugins-openai",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "httpx",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
